Expose card env to ember component cards
My understanding of mobiledoc-kit's Card API is that `env` is part of the public API that a card receives from mobiledoc. If so, then it should be accessible to ember component cards.

An example of a property accessible to cards that wasn't accessible to ember component cards is `env.isInEditor`.