Fallback to parent coords in `tether-to-selection` (#141)
A user reported that the link prompt was always offscreen. I traced this down to `range.getBoundingClient()` returning zero for dimensions and coords in certain cases—I was able to duplicate most reliably by opening a blank editor, hitting “enter”, hitting the “up” key, and finally hitting the link button.

Cases that trigger the issue seemed more prevalent in release Safari than in release Chrome for whatever reason. I could not find the root “cause” of the zeroed-out attributes, but found that falling back to `startContainer` works fine from a UI perspective.